move MXBean to separate package
